requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF @ https://github.com/wtforms/flask-wtf/archive/main.tar.gz
greenlet==3.0.3
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.1
